Fix compilation for linux

+++ source/tests/glkernel-test/tkernel_test.cpp
@@ -18,16 +18,16 @@ TEST_F(tkernel_test, tkernel_indexed_value_access)
 {
     auto fkernel = glkernel::kernel1(2, 4, 8);
 
-    EXPECT_EQ(64, fkernel.size());
-    EXPECT_EQ( 2, fkernel.width());
-    EXPECT_EQ( 4, fkernel.height());
-    EXPECT_EQ( 8, fkernel.depth());
+    EXPECT_EQ(64u, fkernel.size());
+    EXPECT_EQ( 2u, fkernel.width());
+    EXPECT_EQ( 4u, fkernel.height());
+    EXPECT_EQ( 8u, fkernel.depth());
 
     // check if a value set in kernel via spatial reference 
     // equals the value at the expected index ...
 
     const auto findex = fkernel.index(1, 2, 3);
-    EXPECT_EQ(2 * 4 * 3 + 2 * 2 + 1, findex);
+    EXPECT_EQ(static_cast<unsigned int>(2 * 4 * 3 + 2 * 2 + 1), findex);
 
     fkernel[findex] = 1.f;
     EXPECT_EQ(1.f, fkernel.value(1, 2, 3));
@@ -37,16 +37,16 @@ TEST_F(tkernel_test, tkernel_data_access)
 {
     auto fkernel = glkernel::kernel1(8, 2, 4);
 
-    EXPECT_EQ(64, fkernel.size());
-    EXPECT_EQ( 8, fkernel.width());
-    EXPECT_EQ( 2, fkernel.height());
-    EXPECT_EQ( 4, fkernel.depth());
+    EXPECT_EQ(64u, fkernel.size());
+    EXPECT_EQ( 8u, fkernel.width());
+    EXPECT_EQ( 2u, fkernel.height());
+    EXPECT_EQ( 4u, fkernel.depth());
 
     // check if a value set in kernel via spatial reference 
     // equals the value at the expected index ...
 
     const auto findex = fkernel.index(2, 1, 3);
-    EXPECT_EQ(8 * 2 * 3 + 8 * 1 + 2, findex);
+    EXPECT_EQ(static_cast<unsigned int>(8 * 2 * 3 + 8 * 1 + 2), findex);
 
     fkernel[findex] = 1.f;
     EXPECT_EQ(1.f, fkernel.data()[findex]);
@@ -56,10 +56,10 @@ TEST_F(tkernel_test, tkernel_reset)
 {
     auto fkernel = glkernel::kernel1(1024);
 
-    EXPECT_EQ(1024, fkernel.size());
-    EXPECT_EQ(1024, fkernel.width());
-    EXPECT_EQ(   1, fkernel.height());
-    EXPECT_EQ(1, fkernel.depth());
+    EXPECT_EQ(1024u, fkernel.size());
+    EXPECT_EQ(1024u, fkernel.width());
+    EXPECT_EQ(   1u, fkernel.height());
+    EXPECT_EQ(   1u, fkernel.depth());
 
     {   auto accum = 0.f; // checksum
         for (size_t i = 0; i < fkernel.size(); accum += fkernel[i++]);
@@ -100,19 +100,19 @@ TEST_F(tkernel_test, tkernel1_defaults)
 {
     const auto fkernel = glkernel::kernel1{};
 
-    EXPECT_EQ(1, fkernel.size());
-    EXPECT_EQ(1, fkernel.width());
-    EXPECT_EQ(1, fkernel.height());
-    EXPECT_EQ(1, fkernel.depth());
+    EXPECT_EQ(1u, fkernel.size());
+    EXPECT_EQ(1u, fkernel.width());
+    EXPECT_EQ(1u, fkernel.height());
+    EXPECT_EQ(1u, fkernel.depth());
      
     EXPECT_EQ(0.f, fkernel.value(0, 0, 0));
  
     const auto dkernel = glkernel::dkernel1{};
 
-    EXPECT_EQ(1, dkernel.size());
-    EXPECT_EQ(1, dkernel.width());
-    EXPECT_EQ(1, dkernel.height());
-    EXPECT_EQ(1, dkernel.depth());
+    EXPECT_EQ(1u, dkernel.size());
+    EXPECT_EQ(1u, dkernel.width());
+    EXPECT_EQ(1u, dkernel.height());
+    EXPECT_EQ(1u, dkernel.depth());
 
     EXPECT_EQ(0.0, dkernel.value(0, 0, 0));
 }
@@ -121,19 +121,19 @@ TEST_F(tkernel_test, tkernel2_defaults)
 {
     const auto fkernel = glkernel::kernel2{};
 
-    EXPECT_EQ(1, fkernel.size());
-    EXPECT_EQ(1, fkernel.width());
-    EXPECT_EQ(1, fkernel.height());
-    EXPECT_EQ(1, fkernel.depth());
+    EXPECT_EQ(1u, fkernel.size());
+    EXPECT_EQ(1u, fkernel.width());
+    EXPECT_EQ(1u, fkernel.height());
+    EXPECT_EQ(1u, fkernel.depth());
 
     EXPECT_EQ(glm::vec2(0.f, 0.f), fkernel.value(0, 0, 0));
 
     const auto dkernel = glkernel::dkernel2{};
 
-    EXPECT_EQ(1, dkernel.size());
-    EXPECT_EQ(1, dkernel.width());
-    EXPECT_EQ(1, dkernel.height());
-    EXPECT_EQ(1, dkernel.depth());
+    EXPECT_EQ(1u, dkernel.size());
+    EXPECT_EQ(1u, dkernel.width());
+    EXPECT_EQ(1u, dkernel.height());
+    EXPECT_EQ(1u, dkernel.depth());
 
     EXPECT_EQ(glm::dvec2(0.0, 0.0), dkernel.value(0, 0, 0));
 }
@@ -142,19 +142,19 @@ TEST_F(tkernel_test, tkernel3_defaults)
 {
     const auto fkernel = glkernel::kernel3{};
 
-    EXPECT_EQ(1, fkernel.size());
-    EXPECT_EQ(1, fkernel.width());
-    EXPECT_EQ(1, fkernel.height());
-    EXPECT_EQ(1, fkernel.depth());
+    EXPECT_EQ(1u, fkernel.size());
+    EXPECT_EQ(1u, fkernel.width());
+    EXPECT_EQ(1u, fkernel.height());
+    EXPECT_EQ(1u, fkernel.depth());
 
     EXPECT_EQ(glm::vec3(0.f, 0.f, 0.f), fkernel.value(0, 0, 0));
 
     const auto dkernel = glkernel::dkernel3{};
 
-    EXPECT_EQ(1, dkernel.size());
-    EXPECT_EQ(1, dkernel.width());
-    EXPECT_EQ(1, dkernel.height());
-    EXPECT_EQ(1, dkernel.depth());
+    EXPECT_EQ(1u, dkernel.size());
+    EXPECT_EQ(1u, dkernel.width());
+    EXPECT_EQ(1u, dkernel.height());
+    EXPECT_EQ(1u, dkernel.depth());
 
     EXPECT_EQ(glm::dvec3(0.0, 0.0, 0.0), dkernel.value(0, 0, 0));
 }
@@ -163,19 +163,19 @@ TEST_F(tkernel_test, tkernel4_defaults)
 {
     const auto fkernel = glkernel::kernel4{};
 
-    EXPECT_EQ(1, fkernel.size());
-    EXPECT_EQ(1, fkernel.width());
-    EXPECT_EQ(1, fkernel.height());
-    EXPECT_EQ(1, fkernel.depth());
+    EXPECT_EQ(1u, fkernel.size());
+    EXPECT_EQ(1u, fkernel.width());
+    EXPECT_EQ(1u, fkernel.height());
+    EXPECT_EQ(1u, fkernel.depth());
 
     EXPECT_EQ(glm::vec4(0.f, 0.f, 0.f, 0.f), fkernel.value(0, 0, 0));
 
     const auto dkernel = glkernel::dkernel4{};
 
-    EXPECT_EQ(1, dkernel.size());
-    EXPECT_EQ(1, dkernel.width());
-    EXPECT_EQ(1, dkernel.height());
-    EXPECT_EQ(1, dkernel.depth());
+    EXPECT_EQ(1u, dkernel.size());
+    EXPECT_EQ(1u, dkernel.width());
+    EXPECT_EQ(1u, dkernel.height());
+    EXPECT_EQ(1u, dkernel.depth());
 
     EXPECT_EQ(glm::dvec4(0.0, 0.0, 0.0, 0.0), dkernel.value(0, 0, 0));
 }
